feat(toxic-healing-charts): Modularize charts and add specialized flower tracking
This commit introduces a major refactoring of the analytics UI for improved maintainability and adds new data points for tracking specialized flower types.

The monolithic `ChartsPanel` component, which managed the state and configuration for all six charts, has been decomposed. Each chart is now its own modular component within a new `src/components/charts/` directory, simplifying future development and reducing complexity. A common `chartOptions.ts` utility has been created to share base configurations and reduce code duplication.

Key Changes:

-   **Refactor**: Extracted `PopulationChart`, `EnvironmentChart`, `EventsChart`, `FlowerTraitsChart`, `BaseEffectsChart`, and `PerformanceChart` into individual components.
-   **Feature**: The simulation engine now tracks and reports the count of "healing" (negative toxicity) and "toxic" flowers. This data is now visualized as new series on the Population Dynamics chart.
-   **Fix(e2e)**: Resolved a Playwright strict mode violation in the `FlowerPanelController` by updating the actor selection logic to click the `.first()` matching element, making the test more robust when multiple actors occupy the same cell.
-   **Docs**: Updated the README and Planning documents to reflect the new analytics capabilities and refactored component structure.
